requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
fastmcp==2.13.1
mcp>=1.10.0
openai>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.1
